TwitchStreamer/requirements.txt
boto3==1.26.87
botocore==1.29.87
ffmpeg==1.4
jmespath==1.0.1
obsws-python==1.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
websocket-client==1.5.1
